agrego al proyecto las exceptios en Turno, Paciente y Odontologo en los metodos eliminar y guardad y algunso loggers

+++ src/main/java/com/dh/clinica/service/impl/OdontologoService.java
@@ -2,8 +2,12 @@ package com.dh.clinica.service.impl;
 
 
 import com.dh.clinica.entity.Odontologo;
+import com.dh.clinica.exception.BadRequestException;
+import com.dh.clinica.exception.ResourceNotFoundException;
 import com.dh.clinica.repository.IOdontologoRepository;
 import com.dh.clinica.service.IOdontologoService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,6 +15,7 @@ import java.util.Optional;
 
 @Service
 public class OdontologoService implements IOdontologoService {
+    private final Logger logger = LoggerFactory.getLogger(OdontologoService.class);
     private IOdontologoRepository iOdontologoRepository;
 
     public OdontologoService(IOdontologoRepository iOdontologoRepository) {
@@ -19,7 +24,20 @@ public class OdontologoService implements IOdontologoService {
 
     @Override
     public Odontologo guardarOdontologo(Odontologo odontologo) {
-        return iOdontologoRepository.save(odontologo);
+        //valido que el numeroMatricula, nombre y apellido no sean nullos o esten vacios
+        if (odontologo.getNumeroMatricula() == null || odontologo.getNumeroMatricula().isEmpty()){
+            throw new BadRequestException("El numero de la matricula del odontologo no puede estar vacio");
+        }
+        if (odontologo.getNombre() == null || odontologo.getNombre().isEmpty()){
+            throw new BadRequestException("El nombre del odontologo no puede estar vacio");
+        }
+        if (odontologo.getApellido() == null || odontologo.getApellido().isEmpty()){
+            throw new BadRequestException("El apellido del odontologo no puede estar vacio");
+        }
+        //si se valida y esta ok, se guarda el odontologo
+        Odontologo odontologoGuardado = iOdontologoRepository.save(odontologo);
+        logger.info("El odontologo " + odontologoGuardado.getId() + " fue guardado");
+        return odontologoGuardado;
     }
 
     @Override
@@ -39,7 +57,13 @@ public class OdontologoService implements IOdontologoService {
 
     @Override
     public void eliminarOdontologo(Integer id) {
-        iOdontologoRepository.deleteById(id);
+        Optional<Odontologo> odontologoEncontrado = iOdontologoRepository.findById(id);
+        if (odontologoEncontrado.isPresent()){
+            iOdontologoRepository.deleteById(id);
+        } else {
+            throw  new ResourceNotFoundException("El odontologo "+ id + " no fue encontrado");
+        }
+
     }
 
     @Override
@@ -54,12 +78,12 @@ public class OdontologoService implements IOdontologoService {
 
     @Override
     public List<Odontologo> buscarPorNombre(String nombre) {
-        return iOdontologoRepository.finfindByNombre(nombre);
+        return iOdontologoRepository.findByNombre(nombre);
     }
 
     @Override
     public List<Odontologo> buscarPorApellido(String apellido) {
-        return iOdontologoRepository.finfindByApellido(apellido);
+        return iOdontologoRepository.findByApellido(apellido);
     }
